.......... [DEV-672] reverted dev branch to r83982

+++ src/zabbix_agent/zbxconf.c
@@ -51,8 +51,8 @@ void	load_aliases(char **lines)
 
 		if (SUCCEED != parse_key(&r) || ':' != *r)
 		{
-			zabbix_log(LOG_LEVEL_CRIT, "cannot add alias \"%s\": invalid character at position %d",
-					*pline, (int)((r - *pline) + 1));
+			zabbix_log(LOG_LEVEL_CRIT, "cannot add alias \"%s\": invalid character at position %ld",
+					*pline, (r - *pline) + 1);
 			exit(EXIT_FAILURE);
 		}
 
@@ -60,8 +60,8 @@ void	load_aliases(char **lines)
 
 		if (SUCCEED != parse_key(&r) || '\0' != *r)
 		{
-			zabbix_log(LOG_LEVEL_CRIT, "cannot add alias \"%s\": invalid character at position %d",
-					*pline, (int)((r - *pline) + 1));
+			zabbix_log(LOG_LEVEL_CRIT, "cannot add alias \"%s\": invalid character at position %ld",
+					*pline, (r - *pline) + 1);
 			exit(EXIT_FAILURE);
 		}
